feat: added unimodules refactor: changed namespace to xyz.juul.snatchapp chore: added postInstall & setup scripts chore: added solidarity checks chore: added script helpers to package.json fix(android): added icon fonts
