+++ packages/app/android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.snatch",
+    package = "xyz.juul.snatchapp",
 )
 
 android_resource(
     name = "res",
-    package = "com.snatch",
+    package = "xyz.juul.snatchapp",
     res = "src/main/res",
 )
