+++ packages/app/android/app/src/main/java/xyz/juul/snatchapp/MainActivity.java
@@ -1,4 +1,4 @@
-package com.snatch;
+package xyz.juul.snatchapp;
 
 import com.facebook.react.ReactActivity;
