+++ packages/app/ios/Podfile
@@ -1,9 +1,11 @@
-require_relative '../node_modules/react-native/scripts/react_native_pods'
-require_relative '../node_modules/@react-native-community/cli-platform-ios/native_modules'
+require File.join(File.dirname(`node --print "require.resolve('react-native/package.json')"`), "scripts/react_native_pods")
+require File.join(File.dirname(`node --print "require.resolve('@react-native-community/cli-platform-ios/package.json')"`), "native_modules")
+require File.join(File.dirname(`node --print "require.resolve('react-native-unimodules/package.json')"`), "cocoapods")
 
-platform :ios, '11.0'
+platform :ios, '12.0'
 
 target 'Snatch' do
+  use_unimodules!
   config = use_native_modules!
 
   use_react_native!(
