+++ packages/app/package.json
@@ -3,15 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
+    "start": "react-native start",
+    "adb": "adb reverse tcp:9090 tcp:9090 && adb reverse tcp:3000 tcp:3000 && adb reverse tcp:9001 tcp:9001 && adb reverse tcp:8081 tcp:8081",
     "android": "react-native run-android",
     "ios": "react-native run-ios",
-    "start": "react-native start",
     "test": "jest",
-    "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
+    "clean": "react-native-clean-project",
+    "clean-all": "npx react-native clean-project-auto",
+    "format": "npm-run-all format:*",
+    "format:js": "prettier --write \"**/*.js\"",
+    "format:json": "prettier --write \"**/*.json\"",
+    "format:md": "prettier --write \"**/*.md\"",
+    "format:ts": "prettier --write \"**/*.ts{,x}\"",
+    "lint": "eslint App app --fix --ext .js,.jsx,.ts,.tsx && yarn format",
+    "patch": "patch-package",
+    "postinstall": "node ./bin/postInstall"
   },
   "dependencies": {
     "react": "17.0.2",
     "react-native": "0.66.1",
+    "react-native-unimodules": "^0.14.8",
     "react-native-vector-icons": "^8.1.0"
   },
   "devDependencies": {
@@ -24,8 +35,13 @@
     "babel-jest": "^26.6.3",
     "eslint": "^7.14.0",
     "jest": "^26.6.3",
+    "jetifier": "^2.0.0",
     "metro-react-native-babel-preset": "^0.66.2",
+    "npm-run-all": "^4.1.5",
+    "patch-package": "^6.4.7",
+    "postinstall-prepare": "^2.0.0",
     "react-test-renderer": "17.0.2",
+    "solidarity": "^3.0.4",
     "typescript": "^3.8.3"
   },
   "resolutions": {
